# database/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "黑马头条自媒体人管理系统 API",
    "description": "黑马头条是一款面向自媒体创作者的内容管理与数据分析平台。核心功能包括：登录系统、图文数据统计、发布文章、内容列表管理、素材管理、粉丝管理。所有枚举字段均使用 TINYINT 数值存储，与数据库保持一致。",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/channels": {
      "get": {
        "tags": ["频道管理"],
        "summary": "频道列表",
        "description": "获取频道列表，支持启用/禁用筛选，按排序权重升序。",
        "security": [],
        "parameters": [
          {
            "name": "isEnabled",
            "in": "query",
            "description": "1=仅启用 0=仅禁用, 不传返回全部",
            "schema": { "type": "integer", "enum": [0, 1] }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["频道管理"],
        "summary": "创建频道",
        "description": "创建新的文章频道。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "409": { "description": "频道名称已存在" }
        }
      }
    },
    "/channels/{id}": {
      "get": {
        "tags": ["频道管理"],
        "summary": "频道详情",
        "security": [],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" }, "description": "频道ID" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "404": { "description": "频道不存在" }
        }
      },
      "put": {
        "tags": ["频道管理"],
        "summary": "更新频道",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "404": { "description": "频道不存在" }
        }
      }
    },
    "/channels/{id}/status": {
      "patch": {
        "tags": ["频道管理"],
        "summary": "设置频道启用状态",
        "description": "启用或禁用一个频道。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["isEnabled"],
                "properties": {
                  "isEnabled": { "type": "integer", "enum": [0, 1], "description": "1=启用 0=禁用" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "操作成功" },
          "404": { "description": "频道不存在" }
        }
      }
    },
...
    "/materials": {
      "get": {
        "tags": ["素材管理"],
        "summary": "素材列表",
        "description": "分页查询当前用户的素材列表，支持全部/收藏切换。",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "isFavorite", "in": "query", "schema": { "type": "integer", "enum": [0, 1] }, "description": "1=仅收藏 0=全部" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MaterialListResponse" }
              }
            }
          }
        }
      }
    },
    "/materials/upload": {
      "post": {
        "tags": ["素材管理"],
        "summary": "上传素材",
        "description": "上传图片素材。jpg/png 格式，单文件 ≤ 2MB (2097152 字节)。",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "图片文件(jpg/png, ≤2MB)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MaterialResponse" }
              }
            }
          },
          "413": { "description": "文件超过2MB" },
          "422": { "description": "文件类型不是 jpg/png" }
        }
      }
    },
    "/materials/{id}": {
      "get": {
        "tags": ["素材管理"],
        "summary": "素材详情",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MaterialResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["素材管理"],
        "summary": "删除素材（软删除）",
        "description": "已被设为文章封面的素材不允许删除（数据库 RESTRICT 约束）。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "删除成功" },
          "409": { "description": "素材已被设为文章封面，无法删除" }
        }
      }
    },
    "/materials/{id}/restore": {
      "patch": {
        "tags": ["素材管理"],
        "summary": "恢复素材",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "恢复成功" }
        }
      }
    },
    "/materials/{id}/favorite": {
      "patch": {
        "tags": ["素材管理"],
        "summary": "切换收藏状态",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["isFavorite"],
                "properties": {
                  "isFavorite": { "type": "integer", "enum": [0, 1], "description": "1=收藏 0=取消收藏" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "操作成功" }
        }
      }
    },
...
    "/fans": {
      "get": {
        "tags": ["粉丝管理"],
        "summary": "粉丝列表",
        "description": "分页查询当前作者的粉丝列表，支持拉黑过滤。",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "isBlocked", "in": "query", "schema": { "type": "integer", "enum": [0, 1] }, "description": "1=仅拉黑 0=仅正常, 不传全部" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FanListResponse" }
              }
            }
          }
        }
      }
    },
    "/fans/{id}/block": {
      "patch": {
        "tags": ["粉丝管理"],
        "summary": "拉黑粉丝",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "拉黑成功" }
        }
      }
    },
    "/fans/{id}/unblock": {
      "patch": {
        "tags": ["粉丝管理"],
        "summary": "取消拉黑",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "取消拉黑成功" }
        }
      }
    },
    "/fan-messages": {
      "get": {
        "tags": ["粉丝管理"],
        "summary": "私信记录",
        "description": "查询与某个粉丝的私信对话历史。",
        "parameters": [
          { "name": "fanId", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" }, "description": "粉丝ID" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FanMessageListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["粉丝管理"],
        "summary": "发送私信",
        "description": "向粉丝发送私信消息。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FanMessageCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FanMessageResponse" }
              }
            }
          }
        }
      }
    },
    "/fan-stats/overview": {
      "get": {
        "tags": ["粉丝统计"],
        "summary": "粉丝概况概览",
        "description": "获取粉丝概况页4项统计卡片：累计粉丝数量、粉丝累计阅读量、粉丝收益、取消关注量。",
        "parameters": [
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FanStatsOverviewResponse" }
              }
            }
          }
        }
      }
    },
    "/fan-stats": {
      "get": {
        "tags": ["粉丝统计"],
        "summary": "粉丝统计列表",
        "description": "分页查询每日粉丝统计数据。",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FanStatsListResponse" }
              }
            }
          }
        }
      }
    },
    "/fan-stats/trend": {
      "get": {
        "tags": ["粉丝统计"],
        "summary": "阅读量小时趋势",
        "description": "获取某一天的粉丝阅读量24小时趋势数据，对应前端折线图。",
        "parameters": [
          { "name": "statDate", "in": "query", "required": true, "schema": { "type": "string", "format": "date" }, "description": "统计日期" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FanTrendResponse" }
              }
            }
          }
        }
      }
    },
    "/fan-stats/portrait": {
      "get": {
        "tags": ["粉丝统计"],
        "summary": "粉丝画像",
        "description": "获取粉丝画像数据，可按维度筛选。dimension 枚举: 0=性别 1=年龄 2=地域 3=终端 4=活跃时间 5=内容偏好。",
        "parameters": [
          { "name": "dimension", "in": "query", "schema": { "type": "integer", "enum": [0, 1, 2, 3, 4, 5] }, "description": "画像维度, 不传返回全部" },
          { "name": "statDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "统计日期, 默认最新一天" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FanPortraitResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "用户登录后获取的 JWT Token"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "success" }
        }
      },
...
      "Channel": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string", "description": "频道名称" },
          "description": { "type": "string", "nullable": true, "description": "频道描述" },
          "sortOrder": { "type": "integer", "description": "排序权重(越小越靠前)" },
          "isEnabled": { "type": "integer", "enum": [0, 1], "description": "1=启用 0=禁用" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ChannelResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Channel" } } }
        ]
      },
      "ChannelListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "list": { "type": "array", "items": { "$ref": "#/components/schemas/Channel" } }
                }
              }
            }
          }
        ]
      },
      "ChannelCreateRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 32, "description": "频道名称(唯一)", "example": "科技" },
          "description": { "type": "string", "maxLength": 255, "nullable": true, "description": "频道描述" },
          "sortOrder": { "type": "integer", "default": 0, "description": "排序权重" }
        }
      },
      "ChannelUpdateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 32 },
          "description": { "type": "string", "maxLength": 255, "nullable": true },
          "sortOrder": { "type": "integer" }
        }
      },
...
      "Material": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userId": { "type": "integer", "format": "int64" },
          "filename": { "type": "string", "description": "原始文件名" },
          "filePath": { "type": "string", "description": "存储路径/URL" },
          "fileSize": { "type": "integer", "description": "文件大小(字节, ≤2097152)" },
          "mimeType": { "type": "string", "enum": ["image/jpeg", "image/png"], "description": "MIME类型" },
          "isFavorite": { "type": "integer", "enum": [0, 1], "description": "1=已收藏 0=未收藏" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "MaterialResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Material" } } }
        ]
      },
      "MaterialListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "list": { "type": "array", "items": { "$ref": "#/components/schemas/Material" } },
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "pageSize": { "type": "integer" }
                }
              }
            }
          }
        ]
      },
...
      "Fan": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "fanName": { "type": "string", "description": "粉丝用户名/昵称" },
          "fanAvatar": { "type": "string", "nullable": true, "description": "粉丝头像URL" },
          "isBlocked": { "type": "integer", "enum": [0, 1], "description": "1=已拉黑 0=正常" },
          "followedAt": { "type": "string", "format": "date-time", "description": "关注时间" }
        }
      },
      "FanListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "list": { "type": "array", "items": { "$ref": "#/components/schemas/Fan" } },
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "pageSize": { "type": "integer" }
                }
              }
            }
          }
        ]
      },
      "FanMessage": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userId": { "type": "integer", "format": "int64", "description": "发送者(自媒体作者)ID" },
          "fanId": { "type": "integer", "format": "int64", "description": "接收者(粉丝)ID" },
          "content": { "type": "string", "description": "消息内容" },
          "createdAt": { "type": "string", "format": "date-time", "description": "发送时间" }
        }
      },
      "FanMessageResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/FanMessage" } } }
        ]
      },
      "FanMessageCreateRequest": {
        "type": "object",
        "required": ["fanId", "content"],
        "properties": {
          "fanId": { "type": "integer", "format": "int64", "description": "接收者粉丝ID" },
          "content": { "type": "string", "description": "消息内容" }
        }
      },
      "FanMessageListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "list": { "type": "array", "items": { "$ref": "#/components/schemas/FanMessage" } },
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "pageSize": { "type": "integer" }
                }
              }
            }
          }
        ]
      },
      "FanStatsOverview": {
        "type": "object",
        "properties": {
          "totalFanCount": { "type": "integer", "description": "累计粉丝总数" },
          "totalFanReadCount": { "type": "integer", "description": "粉丝累计阅读量" },
          "totalFanRevenue": { "type": "number", "format": "double", "description": "粉丝收益(元)" },
          "totalUnfollowCount": { "type": "integer", "description": "取消关注量" }
        }
      },
      "FanStatsOverviewResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/FanStatsOverview" } } }
        ]
      },
      "FanStatsDailyItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "statDate": { "type": "string", "format": "date" },
          "fanCount": { "type": "integer", "description": "粉丝数量" },
          "fanReadCount": { "type": "integer", "description": "粉丝阅读量" },
          "fanRevenue": { "type": "number", "format": "double", "description": "粉丝收益(元)" },
          "unfollowCount": { "type": "integer", "description": "取消关注量" },
          "newFollowCount": { "type": "integer", "description": "新增关注量" }
        }
      },
      "FanStatsListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "list": { "type": "array", "items": { "$ref": "#/components/schemas/FanStatsDailyItem" } },
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "pageSize": { "type": "integer" }
                }
              }
            }
          }
        ]
      },
      "TrendHourItem": {
        "type": "object",
        "properties": {
          "hour": { "type": "integer", "minimum": 0, "maximum": 23, "description": "小时(0-23)" },
          "readCount": { "type": "integer", "description": "该小时阅读量" }
        }
      },
      "FanTrendResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "statDate": { "type": "string", "format": "date" },
                  "hours": { "type": "array", "items": { "$ref": "#/components/schemas/TrendHourItem" } }
                }
              }
            }
          }
        ]
      },
      "PortraitItem": {
        "type": "object",
        "properties": {
          "dimensionKey": { "type": "string", "description": "维度具体值, 如 male / 18-23 / 广东 / iOS" },
          "dimensionKeyLabel": { "type": "string", "description": "维度值中文标签" },
          "dimensionValue": { "type": "integer", "description": "数量" },
          "percentage": { "type": "number", "format": "double", "nullable": true, "description": "占比(%)" }
        }
      },
      "PortraitDimensionData": {
        "type": "object",
        "properties": {
          "dimension": { "type": "integer", "enum": [0, 1, 2, 3, 4, 5], "description": "0=性别 1=年龄 2=地域 3=终端 4=活跃时间 5=内容偏好" },
          "dimensionLabel": { "type": "string", "description": "维度中文名" },
          "chartType": { "type": "string", "enum": ["doughnut", "bar", "map"], "description": "前端图表类型" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/PortraitItem" } }
        }
      },
      "FanPortraitResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "description": "传dimension参数时返回单个维度, 不传返回全部维度",
                "properties": {
                  "statDate": { "type": "string", "format": "date" },
                  "dimension": { "type": "integer", "enum": [0, 1, 2, 3, 4, 5] },
                  "dimensionLabel": { "type": "string" },
                  "chartType": { "type": "string", "enum": ["doughnut", "bar", "map"] },
                  "items": { "type": "array", "items": { "$ref": "#/components/schemas/PortraitItem" } },
                  "portraits": { "type": "array", "items": { "$ref": "#/components/schemas/PortraitDimensionData" } }
                }
              }
            }
          }
        ]
      },
...
    }
  }
}